Initial commit
Huge initial commit as I didn't expect this to get complicated and look this nice heh.
Contains the following:
Good UI
Backend that indexes new files almost immediately and is very performant overall
GPU detection which is essentially distributed computing as my NAS where the UI and indexing runs doesn't have a GPU for the AI detection logic so this service runs in conjunction with the UI/backend monorepo. By default the backend drops the camera jpg into a shared network folder where the gpu detection logic picks up the file and runs it through an image classification model. That model spits out json and the jpg with boxes around classifed things. The backend waits for that, picks it up and then ingests both of those into the DB updating the UI.